+++ packages/desktop/features/governance.features.ts
@@ -1,0 +1,5 @@
+const governanceFeatures = {
+    enabled: true,
+}
+
+export default governanceFeatures
