+++ packages/shared/lib/contexts/governance/interfaces/organization.interface.ts
