+++ packages/shared/lib/contexts/governance/interfaces/proposal.interface.ts
