+++ packages/shared/lib/contexts/governance/stores/selected-proposal.store.ts
@@ -1,4 +1,4 @@
 import { writable } from 'svelte/store'
-import { IProposal } from '@core/governance/interfaces'
+import { IProposal } from '@contexts/governance/interfaces'
 
 export const selectedProposal = writable<IProposal>(null)
