+++ packages/shared/lib/core/profile-manager/api/registerParticipationEvent.ts
@@ -1,7 +1,10 @@
-import { addProposalState } from '@core/governance'
-import { profileManager } from '@core/profile-manager/stores'
-import type { Event, EventId, Node } from '@iota/wallet'
 import { get } from 'svelte/store'
+
+import type { Event, EventId, Node } from '@iota/wallet'
+
+import { addProposalState } from '@contexts/governance/stores'
+
+import { profileManager } from '../stores'
 
 export async function registerParticipationEvent(eventId: EventId, nodes: Node[]): Promise<Event> {
     const manager = get(profileManager)
